refactor: remove app dir and ajust scripts
