+++ __tests__/unit/mocks/validator/validator-mock.ts
@@ -1,4 +1,4 @@
-import { IValidator } from '../../../../src/app/presentation/protocols/validator';
+import { IValidator } from '../../../../src/presentation/protocols/validator';
 
 export class ValidatorMock implements IValidator {
   validate(input: any): Promise<string[] | {}> | string[] | {} {
