+++ package.json
@@ -8,8 +8,8 @@
   },
   "scripts": {
     "build": "tsc",
-    "start:prod": "node ./build/app/main/server.js",
-    "start:dev": "npx ts-node-dev --respawn --transpile-only --inspect --ignore-watch node_modules src/app/main/server.ts",
+    "start:prod": "node ./build/main/server.js",
+    "start:dev": "npx ts-node-dev --respawn --transpile-only --inspect --ignore-watch node_modules src/main/server.ts",
     "test": "npx jest --passWithNoTests --silent --noStackTrace --runInBand",
     "test:unit": "npm run test -- --watch -c jest-unit.config.js",
     "test:integration": "yarn test -- --watch -c jest-integration.config.js",
