+++ src/domain/entities/pokemon-characteristics.ts
