+++ src/domain/entities/pokemon.ts
