+++ src/domain/entities/power-status.ts
