+++ src/domain/entities/type.ts
