+++ src/domain/entities/weather.ts
