+++ src/domain/util/pagination-data.ts
